moved stuff into models and utils folders and deleted unused WeatherResult.kt
