--- app/src/main/java/com/tstudioz/fax/fme/feature/home/WeatherResult.kt
@@ -1,6 +1,0 @@
-package com.tstudioz.fax.fme.feature.home
-
-sealed class WeatherResult<out T : Any> {
-    data class Success<out T : Any>(val body: String) : WeatherResult<T>()
-    class Failure(exception: Throwable) : WeatherResult<Nothing>()
-}
