+++ app/src/main/java/com/tstudioz/fax/fme/feature/home/repository/WeatherRepositoryInterface.kt
@@ -1,6 +1,6 @@
 package com.tstudioz.fax.fme.feature.home.repository
 
-import com.tstudioz.fax.fme.feature.home.WeatherDisplay
+import com.tstudioz.fax.fme.feature.home.models.WeatherDisplay
 
 interface WeatherRepositoryInterface {
